// src/providers/docProvider/models/docFields.ts
export enum DocField {
  PERIODICITY = 'PERIODICITY',
  AMOUNT_CHARGES = 'AMOUNT_CHARGES',
  CHARGED_EVERY_X_DAYS = 'CHARGED_EVERY_X_DAYS',
  START_DATE = 'START_DATE',
  STATUS = 'STATUS',
  STATUS_DATE = 'STATUS_DATE',
  CANCELLATION_DATE = 'CANCELLATION_DATE',
  VALUE = 'VALUE',
  NEXT_CYCLE = 'NEXT_CYCLE',
  SUBSCRIBER_ID = 'SUBSCRIBER_ID',
  NA = 'N/A',
}